requirements.txt
flask
tensorflow-cpu==2.6.0
pillow
numpy